renamed DelegateProfile page, removed missing data, consolidating validator and candidate profiles,
