+++ app/src/renderer/components/staking/PageDelegateProfile.vue
@@ -1,0 +1,57 @@
+<template lang="pug">
+page(:title="delegateType + ' Profile'")
+  div(slot="menu"): tool-bar
+    router-link(to='/staking')
+      i.material-icons arrow_back
+      .label Back
+
+  part(:title="delegateType + ' Info'")
+    list-item(dt='Moniker' :dd='delegate.moniker')
+    list-item(dt='Website' :dd='delegate.website')
+    list-item(dt='Address' :dd='delegate.owner.addr')
+    list-item(dt='Public Key' :dd='delegate.pub_key.data')
+
+  part(:title="delegateType + ' Description'")
+    text-block(:content="delegate.details || 'No description available.'")
+
+  part(:title="delegateType + ' Details'" v-if="isValidator")
+    list-item(dt="Total Vote Power" :dd="delegate.voting_power"
+      :to="{ name: 'validator-power', params: { validator: this.$route.params.delegate }}")
+</template>
+
+<script>
+import { mapGetters } from 'vuex'
+import Btn from '@nylira/vue-button'
+import LiCopy from 'common/NiLiCopy'
+import ListItem from 'common/NiListItem'
+import Page from 'common/NiPage'
+import Part from 'common/NiPart'
+import TextBlock from 'common/TextBlock'
+import ToolBar from 'common/NiToolBar'
+export default {
+  name: 'page-delegate',
+  components: {
+    Btn,
+    LiCopy,
+    ListItem,
+    Page,
+    Part,
+    TextBlock,
+    ToolBar
+  },
+  computed: {
+    ...mapGetters(['delegates']),
+    delegate () {
+      if (this.delegates.delegates && this.$route.params.delegate) {
+        return this.delegates.delegates.find(v => v.id === this.$route.params.delegate)
+      }
+    },
+    isValidator () {
+      return this.delegate.voting_power > 0
+    },
+    delegateType () {
+      return this.isValidator ? 'Validator' : 'Candidate'
+    }
+  }
+}
+</script>
