+++ test/unit/specs/components/staking/PageDelegates.spec.js
@@ -1,6 +1,6 @@
 import setup from '../../../helpers/vuex-setup'
 import htmlBeautify from 'html-beautify'
-import PageDelegates from 'renderer/components/staking/PageDelegates'
+import PageDelegates from 'renderer/components/staking/PageDelegatesProfile'
 
 describe('PageDelegates', () => {
   let wrapper, store
